requirements.txt
discord.py>=2.3.0
yt-dlp>=2023.7.6
spotipy>=2.23.0
python-dotenv>=1.0.0
mysql-connector-python>=8.1.0
aiohttp>=3.8.0
instaloader>=4.9.0
requests>=2.31.0